requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asyncpg==0.30.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
